requirements.txt
flask
requests
RPi.GPIO; platform_system=='Linux'
